fixing circular import for gps utilities
