+++ src/util/gps_utils.py
@@ -1,0 +1,38 @@
+import numpy as np
+import pymap3d
+
+# import waypoint
+from mrover.msg import Waypoint, WaypointType, GPSWaypoint
+from util.SE3 import SE3
+import rospy
+
+REF_LAT = rospy.get_param("gps_linearization/reference_point_latitude")
+REF_LON = rospy.get_param("gps_linearization/reference_point_longitude")
+
+
+def convert_gps_to_cartesian(waypoint: GPSWaypoint) -> Waypoint:
+    """
+    Converts a GPSWaypoint into a "Waypoint" used for publishing to the CourseService.
+    """
+    # Create odom position based on GPS latitude and longitude
+    odom = np.array(
+        pymap3d.geodetic2enu(
+            waypoint.latitude_degrees, waypoint.longitude_degrees, 0.0, REF_LAT, REF_LON, 0.0, deg=True
+        )
+    )
+    # zero the z-coordinate of the odom because even though the altitudes are set to zero,
+    # two points on a sphere are not going to have the same z coordinate
+    # navigation algorithmns currently require all coordinates to have zero as the z coordinate
+    odom[2] = 0
+
+    return Waypoint(fiducial_id=waypoint.id, tf_id=f"course{waypoint.id}", type=waypoint.type), SE3(position=odom)
+
+
+def convert_cartesian_to_gps(coordinate: np.ndarray) -> GPSWaypoint:
+    """
+    Converts a coordinate to a GPSWaypoint (used for sending data back to basestation)
+    """
+    lat, lon, _ = pymap3d.enu2geodetic(
+        e=coordinate[0], n=coordinate[1], u=0.0, lat0=REF_LAT, lon0=REF_LON, h0=0.0, deg=True
+    )
+    return GPSWaypoint(lat, lon, WaypointType(val=WaypointType.NO_SEARCH), 0)
